MAGETWO-55858: Output escaping methods shouldn't be part of AbstractBlock

+++ lib/internal/Magento/Framework/View/Element/AbstractBlock.php
@@ -877,7 +877,7 @@ abstract class AbstractBlock extends \Magento\Framework\DataObject implements Bl
         $namespace = substr(
             $className,
             0,
-            strpos($className, '\\' . 'Block' . '\\')
+            (int)strpos($className, '\\' . 'Block' . '\\')
         );
         return str_replace('\\', '_', $namespace);
     }
